Moved settings navigation into a modal stack
