+++ App/Navigation/AppNavigation/Modal/Primary/Drawer/Groups/index.tsx
@@ -1,16 +1,15 @@
 import { createStackNavigator } from 'react-navigation'
 
-import ThreadsList from '../../../../../Containers/ThreadsList'
-import ThreadsManager from '../../../../../Containers/ThreadsManager'
+import Groups from '../../../../../../Containers/Groups'
 
-import styles, { headerTintColor } from '../../../../Styles/NavigationStyles'
+import styles, { headerTintColor } from '../../../../../Styles/NavigationStyles'
 
 const nav = createStackNavigator(
   {
-    SharedPhotos: ThreadsList,
-    ThreadsManager
+    Groups
   },
   {
+    headerMode: 'float',
     defaultNavigationOptions: {
       headerStyle: styles.header,
       headerTitleStyle: styles.headerTitle,
