--- App/Navigation/AppNavigation/Primary/Drawer/Groups/index.tsx
@@ -1,21 +1,0 @@
-import { createStackNavigator } from 'react-navigation'
-
-import Groups from '../../../../../Containers/Groups'
-
-import styles, { headerTintColor } from '../../../../Styles/NavigationStyles'
-
-const nav = createStackNavigator(
-  {
-    Groups
-  },
-  {
-    headerMode: 'float',
-    defaultNavigationOptions: {
-      headerStyle: styles.header,
-      headerTitleStyle: styles.headerTitle,
-      headerTintColor
-    }
-  }
-)
-
-export default nav
