--- App/Navigation/AppNavigation/Primary/Drawer/Wallet/index.ts
@@ -1,25 +1,0 @@
-import { createStackNavigator } from 'react-navigation'
-
-import Wallet from '../../../../../Containers/Wallet'
-import PrivatePhotoDetail from '../../../../../SB/views/PhotoDetail'
-import AddCaptionScreen from '../../../../../Containers/AddCaptionScreen'
-
-import styles, {headerTintColor} from '../../../../Styles/NavigationStyles'
-
-const nav = createStackNavigator(
-  {
-    Wallet,
-    PrivatePhotoDetail,
-    WalletSharePhoto: AddCaptionScreen
-  },
-  {
-    headerMode: 'float',
-    defaultNavigationOptions: {
-      headerStyle: styles.header,
-      headerTitleStyle: styles.headerTitle,
-      headerTintColor
-    }
-  }
-)
-
-export default nav
